+++ cmd/zed/dev/vcache/project/command.go
@@ -17,13 +17,13 @@ import (
 var Project = &charm.Spec{
 	Name:  "project",
 	Usage: "project [flags] field[,field...] path",
-	Short: "read a ZST file and run a projection as a test",
+	Short: "read a VNG file and run a projection as a test",
 	Long: `
-The project command reads ZST vectors from
-a ZST storage objects (local files or s3 objects) and outputs
+The project command reads VNG vectors from
+a VNG storage objects (local files or s3 objects) and outputs
 the reconstructed ZNG row data as a projection of one or more fields.
 
-This command is most useful for testing the ZST vector cache.
+This command is most useful for testing the VNG vector cache.
 `,
 	New: newCommand,
 }
@@ -50,7 +50,7 @@ func (c *Command) Run(args []string) error {
 	}
 	defer cleanup()
 	if len(args) < 2 {
-		return errors.New("zst read: must be run with a single path argument followed by one or more fields")
+		return errors.New("VNG read: must be run with a single path argument followed by one or more fields")
 	}
 	uri, err := storage.ParseURI(args[0])
 	if err != nil {