+++ lake/data/writer_test.go
@@ -9,7 +9,7 @@ import (
 	"github.com/brimdata/zed/order"
 	"github.com/brimdata/zed/pkg/field"
 	"github.com/brimdata/zed/pkg/storage"
-	"github.com/brimdata/zed/zio/zstio"
+	"github.com/brimdata/zed/zio/vngio"
 	"github.com/brimdata/zed/zson"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -28,10 +28,10 @@ func TestDataReaderWriterVector(t *testing.T) {
 	require.NoError(t, w.Write(zson.MustParseValue(zctx, "{a:3,b:6}")))
 	require.NoError(t, w.Close(ctx))
 	require.NoError(t, data.CreateVector(ctx, engine, tmp, object.ID))
-	// Read back the ZST file and make sure it's the same.
+	// Read back the VNG file and make sure it's the same.
 	get, err := engine.Get(ctx, object.VectorURI(tmp))
 	require.NoError(t, err)
-	reader, err := zstio.NewReader(zed.NewContext(), get)
+	reader, err := vngio.NewReader(zed.NewContext(), get)
 	require.NoError(t, err)
 	v, err := reader.Read()
 	require.NoError(t, err)