+++ vng/vector/io.go
@@ -1,7 +1,7 @@
 // Package vector implements the organization of Zed data on storage as
-// vectors in a ZST vector storage object.
+// vectors in a VNG vector storage object.
 //
-// A ZST object is created by allocating a Writer for any top-level Zed type
+// A VNG object is created by allocating a Writer for any top-level Zed type
 // via NewWriter.  The object to be written is wrapped
 // in a Spiller with a vector threshold.  Output is streamed to the underlying spiller
 // in a single pass.
@@ -13,14 +13,14 @@
 //
 // After all of the Zed data is written, a metadata section is written consisting
 // of segment maps for each vector, each obtained by calling the Metadata
-// method on the zst.Writer interface.
+// method on the vng.Writer interface.
 //
 // Nulls for complex types are encoded by a special Nulls object.  Each complex
 // type is wrapped by a NullsWriter, which runlength encodes any alternating
 // sequences of nulls and values.  If no nulls are encountered, then the Nulls
 // object is omitted from the metadata.
 //
-// Data is read from a ZST file by scanning the metadata maps to build
+// Data is read from a VNG file by scanning the metadata maps to build
 // vector Readers for each Zed type by calling NewReader with the metadata, which
 // recusirvely builds reassembly segments.  An io.ReaderAt is passed to NewReader
 // so each vector reader can access the underlying storage object and read its
@@ -49,10 +49,10 @@ type Writer interface {
 	Write(zcode.Bytes) error
 	// Push all in-memory vector data to the storage layer.
 	Flush(bool) error
-	// Metadata returns the data structure conforming to the ZST specification
+	// Metadata returns the data structure conforming to the VNG specification
 	// describing the layout of vectors.  This is called after all data is
 	// written and flushed by the Writer with the result marshaled to build
-	// the metadata section of the ZST file.
+	// the metadata section of the VNG file.
 	Metadata() Metadata
 }
 
@@ -74,7 +74,7 @@ func NewWriter(typ zed.Type, spiller *Spiller) Writer {
 		return NewNullsWriter(NewUnionWriter(typ, spiller), spiller)
 	default:
 		if !zed.IsPrimitiveType(typ) {
-			panic(fmt.Sprintf("unsupported type in ZST file: %T", typ))
+			panic(fmt.Sprintf("unsupported type in VNG file: %T", typ))
 		}
 		return NewPrimitiveWriter(typ, spiller)
 	}
@@ -118,6 +118,6 @@ func NewReader(meta Metadata, r io.ReaderAt) (Reader, error) {
 	case *Primitive:
 		return NewPrimitiveReader(meta, r), nil
 	default:
-		return nil, fmt.Errorf("unknown ZST metadata type: %T", meta)
+		return nil, fmt.Errorf("unknown VNG metadata type: %T", meta)
 	}
 }