+++ zson/parser-values.go
@@ -5,7 +5,6 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"net"
 	"net/netip"
 	"strconv"
 	"time"
@@ -201,7 +200,7 @@ func (p *Parser) matchPrimitive() (*astzed.Primitive, error) {
 		typ = "duration"
 	} else if _, err := netip.ParsePrefix(s); err == nil {
 		typ = "net"
-	} else if ip := net.ParseIP(s); ip != nil {
+	} else if _, err := netip.ParseAddr(s); err == nil {
 		typ = "ip"
 	} else if len(s) >= 2 && s[0:2] == "0x" {
 		if len(s) == 2 {